apply gates access control in controllers

+++ app/Http/Controllers/TaskController.php
@@ -8,6 +8,7 @@ use App\Models\Task;
 use App\Repositories\Contracts\TaskRepository;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -15,6 +16,8 @@ class TaskController extends Controller
 
     public function store(Project $project, StoreUpdateTaskRequest $request)
     {
+        Gate::authorize('has-permission', 'create task');
+
         $task = $this->taskRepository->create($project, $request->validated());
 
         return new JsonResponse($task, Response::HTTP_CREATED);
@@ -22,6 +25,8 @@ class TaskController extends Controller
 
     public function update(StoreUpdateTaskRequest $request, Task $task)
     {
+        Gate::authorize('has-permission', 'edit task');
+
         $this->taskRepository->update($task, $request->validated());
 
         return new JsonResponse($task->fresh(), Response::HTTP_OK);
@@ -29,6 +34,8 @@ class TaskController extends Controller
 
     public function destroy(Task $task)
     {
+        Gate::authorize('has-permission', 'delete task');
+
         $this->taskRepository->delete($task);
 
         return new JsonResponse(status: Response::HTTP_NO_CONTENT);
